+++ eras/shelley/test-suite/bench/BenchValidation.hs
@@ -5,8 +5,10 @@
 {-# LANGUAGE GADTs #-}
 {-# LANGUAGE RankNTypes #-}
 {-# LANGUAGE ScopedTypeVariables #-}
+{-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE TypeOperators #-}
+{-# LANGUAGE UndecidableInstances #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
 {-# OPTIONS_GHC -Wno-unused-binds #-}
 
@@ -17,6 +19,7 @@ module BenchValidation (
   benchreValidate,
   applyBlock,
   sizes,
+  UpdateInputs,
   updateChain,
   updateAndTickChain,
   genUpdateInputs,
@@ -38,9 +41,6 @@ import Cardano.Protocol.Crypto
 import Cardano.Protocol.TPraos.API (
   ChainDepState (..),
   ChainTransitionError,
-  GetLedgerView,
-  LedgerView (..),
-  currentLedgerView,
   tickChainDepState,
   updateChainDepState,
  )
@@ -77,7 +77,7 @@ validateInput ::
   , EraRule "LEDGERS" era ~ API.ShelleyLEDGERS era
   , QC.HasTrace (API.ShelleyLEDGERS era) (GenEnv MockCrypto era)
   , API.ApplyBlock era
-  , GetLedgerView era
+  , API.ShelleyEraForecast era
   , MinLEDGER_STS era
   ) =>
   Int ->
@@ -123,21 +123,16 @@ benchreValidate ::
 benchreValidate (ValidateInput globals state (Block bh txs)) =
   API.applyBlockNoValidaton globals state (Block (makeHeaderView bh Nothing) txs)
 
--- ==============================================================
-
-data UpdateInputs
+data UpdateInputs era
   = UpdateInputs
       !Globals
-      !LedgerView
+      !(API.Forecast API.Current era)
       !(BHeader MockCrypto)
       !ChainDepState
 
-instance Show UpdateInputs where
+instance Show (API.Forecast API.Current era) => Show (UpdateInputs era) where
   show (UpdateInputs _globals vl bh st) =
     show vl ++ "\n" ++ show bh ++ "\n" ++ show st
-
-instance NFData LedgerView where
-  rnf (LedgerView _D _extraEntropy _pool _delegs _ccd) = ()
 
 -- TODO: move upstream
 instance Crypto c => NFData (BHeader c) where
@@ -149,7 +144,7 @@ instance NFData ChainDepState where
 instance NFData (ChainTransitionError c) where
   rnf _ = ()
 
-instance NFData UpdateInputs where
+instance NFData (API.Forecast API.Current era) => NFData (UpdateInputs era) where
   rnf (UpdateInputs g lv bh st) =
     seq (rnf g) (seq (rnf lv) (seq (rnf bh) (rnf st)))
 
@@ -158,18 +153,18 @@ genUpdateInputs ::
   ( EraGen era
   , EraStake era
   , MinLEDGER_STS era
-  , GetLedgerView era
+  , API.ShelleyEraForecast era
   , EraRule "LEDGERS" era ~ API.ShelleyLEDGERS era
   , QC.HasTrace (API.ShelleyLEDGERS era) (GenEnv MockCrypto era)
   , API.ApplyBlock era
   ) =>
   Int ->
-  IO UpdateInputs
+  IO (UpdateInputs era)
 genUpdateInputs utxoSize = do
   let ge = genEnv (Proxy :: Proxy era) defaultConstants
   chainstate <- genChainState utxoSize ge
   (Block blockheader _) <- genBlock ge chainstate
-  let ledgerview = currentLedgerView (chainNes chainstate)
+  let forecast = API.currentForecast (chainNes chainstate)
   let (ChainState _newepochState keys eta0 etaV etaC etaH slot) = chainstate
   let prtclState = PrtclState keys eta0 etaV
   let ticknState = TicknState etaC etaH
@@ -179,20 +174,21 @@ genUpdateInputs utxoSize = do
   pure
     ( UpdateInputs
         testGlobals
-        ledgerview
+        forecast
         blockheader
         (ChainDepState prtclState ticknState nonce)
     )
 
 updateChain ::
-  UpdateInputs ->
+  API.ShelleyEraForecast era =>
+  UpdateInputs era ->
   Either (ChainTransitionError MockCrypto) ChainDepState
-updateChain (UpdateInputs gl lv bh st) = updateChainDepState gl lv bh st
+updateChain (UpdateInputs gl lv bh st) = updateChainDepState @API.Current gl lv bh st
 
 updateAndTickChain ::
-  UpdateInputs ->
+  API.ShelleyEraForecast era =>
+  UpdateInputs era ->
   Either (ChainTransitionError MockCrypto) ChainDepState
 updateAndTickChain (UpdateInputs gl lv bh st) =
-  updateChainDepState gl lv bh
-    . tickChainDepState gl lv True
-    $ st
+  updateChainDepState @API.Current gl lv bh $
+    tickChainDepState @API.Current gl lv True st