+++ eras/shelley/impl/CHANGELOG.md
@@ -6,6 +6,18 @@
 
 ## 1.18.0.0
 
+* Add `Shelley.API.Forecast` and `Shelley.Forecast`:
+  - Add `EraForecast` and `ShelleyEraForecast` typeclasses to deprecate `GetLedgerView` from `cardano-ledger-tpraos`.
+  - Add `currentForecast` and `futureForecast` functions to deprecate `currentLedgerView` and `futureLedgerView`.
+  - Add `ShelleyForecast` to deprecate `LedgerView` for TPraos.
+    + `mkShelleyForecast`
+    + `sfPoolDistrL`
+    + `sfMaxBlockHeaderSizeL`
+    + `sfMaxBlockBodySizeL`
+    + `sfProtocolVersionL`
+    + `sfGenDelegsL`
+    + `sfDecentralizationL`
+    + `sfExtraEntropyL`
 * Subsume delegations into `ActiveStake`.
   - `epochStateStakeDistrL` returns `ActiveStake`
   - Remove `fvDelegs` from `FreeVars`.